requirements.txt
-e .

# For testing
pytest==7.2.1
pytest-dotenv==0.5.2

Flask-Cors==3.0.10


# third-party (libraries)
rake_nltk==1.0.6
ipython==8.10.0

# linting stubs
types-requests==2.28.11.8
types-setuptools==67.1.0.0

# linting
black==22.12.0
mypy==0.991
pre-commit==3.2.0
pylint==2.15.10
ruff==0.0.259
types-redis==4.5.5.0